Generic Stack
Templated the Stack structure
Moved Stack::is to CallStack as Stack is now a purely "stack" focused
structure

+++ Interpreter/WIP/CallStack.cpp
@@ -24,7 +24,7 @@ namespace dust {
 	}
 
 	void CallStack::settop(int idx) {
-		idx = std::max(normalize(idx), 0);
+		idx = (size_t)std::max(normalize(idx), 0);
 
 		if (idx > size())
 			while (size() < idx)
